% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyT1BC.R
\name{classifyT1BC}
\alias{classifyT1BC}
\title{Classify T1 bladder cancer samples}
\usage{
classifyT1BC(gexp, id_type = c("symbol", "ensembl_id", "entrez_id")[1])
}
\arguments{
\item{gexp}{a matrix or data.frame with rows as genes and columns as samples.}

\item{id_type}{type of gene annation used in the expression matrix. One of:
"symbol", "ensembl_id" or "entrez_id" for HGNC symbol, ENSEMBL gene ID, and
Entrez Gene ID, respectively.}
}
\value{
A data.frame with `ncol(gexp)` rows and 8 columns:
* `id` contains the colnames of the original expression matrix
* `class` contains the class prediction for the sample
* `class_prob` shows the model probability for the assigned class
* `cl_1_prob` to `cl_2_prob` contain the probabilities for each class
}
\description{
This samples provides class calls and class probabilities given log-transformed
normalized expression data. The expression data must be in the form of a
matrix of data.frame, where rows are genes and columns are samples.
}
